requirements.txt
nicovideo.py